bin: split pipeline context support
